展示小区详情和小区列表08

+++ core-model/src/main/java/com/toutiao/web/domain/query/VillageResponse.java
@@ -65,8 +65,8 @@ public class VillageResponse {
     private String decorationType;  //装修标准
     private String ringRoadName;  //详情环线
     private Integer ringRoadDistance;  //详情环线距离
-    private Integer tongbi;  //同比
-    private Integer huanbi;  //环比
+    private Double tongbi;  //同比
+    private Double huanbi;  //环比
 //    private Double pm;  //pm2.5
 //    private Double noiseindex;  //噪声指数
 //    private String[] mainpushhouse;  //主推房源
